+++ lib/src/util/network_util.dart
@@ -4,9 +4,7 @@ import 'dart:io';
 import 'package:dawnn_client/src/util/client_util.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_image_compress/flutter_image_compress.dart';
 import 'package:http/http.dart' as http;
-import 'package:path_provider/path_provider.dart';
 import 'package:top_snackbar_flutter/custom_snack_bar.dart';
 import 'package:top_snackbar_flutter/top_snack_bar.dart';
 
@@ -18,7 +16,7 @@ class NetworkUtils {
   /// Posts the image at [imagePath] to Dawnn server.
   static void postImage(BuildContext context, String imagePath) async {
     // This is kind of a long method...
-    var data = Data(await _compressToBase64(File(imagePath)),
+    var data = Data(await ClientUtils.compressToBase64(File(imagePath)),
         await ClientUtils.getHWID(context), await ClientUtils.getLocation());
     var body = json.encode(data.toJson());
 
@@ -66,31 +64,8 @@ class NetworkUtils {
         context, CustomSnackBar.info(message: message, backgroundColor: color));
   }
 
-  /// Compress an image to 95% of its original quality,
-  /// correct angle, and convert the image [file] to the base64
-  /// format.
-  static Future<String> _compressToBase64(File file) async {
-    var dir = await getTemporaryDirectory();
-    var targetPath = dir.absolute.path + "/temp.jpg";
-    File result = await FlutterImageCompress.compressAndGetFile(
-        file.absolute.path, targetPath,
-        autoCorrectionAngle: true);
-    final bytes = await result.readAsBytes();
-    return base64Encode(bytes);
-  }
-
-  /// Decode a base64 image [source] to a file.
-  /// To be used whenever we implement
-  /// getting images from dawn server.
-  static File _fromBase64(String source) {
-    var bytes = base64.decode(source);
-    var file = File("newImage.jpg");
-    file.writeAsBytesSync(bytes);
-    return file;
-  }
-
-  // TODO How to elegantly get image from server?
-  static Future<String> getImage() async {
+  /// Get images as base64 strings in a list
+  static Future<String> getImages() async {
     // Implement after team discussion.
     // For now, use placeholder image.
 
@@ -102,6 +77,7 @@ class NetworkUtils {
       port: 2423,
     ));
 
+    // TODO Handle errors
     client.close();
     return response.body;
   }
